Functions/ProcessMessage/requirements.txt
adal==1.0.2
applicationinsights==0.11.7
argcomplete==1.9.4
asn1crypto==0.24.0
azure-cli-core==2.0.40
azure-cli-nspkg==3.0.3
azure-common==1.1.16
azure-graphrbac==0.53.0
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.51.1
azure-mgmt-batch==5.0.1
azure-mgmt-cdn==3.0.0
azure-mgmt-cognitiveservices==3.0.0
azure-mgmt-compute==4.4.0
azure-mgmt-containerservice==4.3.0
azure-mgmt-cosmosdb==0.5.2
azure-mgmt-datafactory==0.6.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-eventgrid==2.0.0rc2
azure-mgmt-iothub==0.7.0
azure-mgmt-keyvault==1.1.0
azure-mgmt-monitor==0.5.2
azure-mgmt-network==2.5.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.2.0
azure-mgmt-redis==5.0.0
azure-mgmt-resource==2.1.0
azure-mgmt-sql==0.11.0
azure-mgmt-storage==3.1.1
azure-mgmt-subscription==0.2.0
azure-mgmt-web==0.40.0
azure-nspkg==3.0.2
azure-storage-blob==1.4.0
azure-storage-common==1.4.0
azure-storage-queue==1.4.0
backports.functools-lru-cache==1.5
bcrypt==3.1.5
boto3==1.9.75
botocore==1.12.75
c7n==0.8.33.1
c7n-azure==0.4
c7n-mailer==0.4.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
colorama==0.4.1
cryptography==2.4.2
datadog==0.26.0
decorator==4.3.0
distlib==0.2.8
docutils==0.14
futures==3.1.1
humanfriendly==4.17
idna==2.8
isodate==0.6.0
Jinja2==2.10
jmespath==0.9.3
jsonpatch==1.23
jsonpointer==2.0
jsonschema==2.6.0
knack==0.3.3
ldap3==2.5.2
MarkupSafe==1.1.0
msrest==0.6.3
msrestazure==0.6.0
oauthlib==3.0.0
paramiko==2.4.2
pyasn1==0.4.5
pycparser==2.19
Pygments==2.3.1
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==18.0.0
python-dateutil==2.7.5
python-http-client==3.1.0
PyYAML==3.13
redis==3.0.1
requests==2.21.0
requests-oauthlib==1.0.0
ruamel.yaml==0.15.42
s3transfer==0.1.13
sendgrid==5.6.0
six==1.12.0
slackclient==1.3.0
tabulate==0.8.2
urllib3==1.24.1
websocket-client==0.54.0
